refactor(ball): ball number를 1급 객체로 사용하도록 변경한다
